addressed filter move to core issue
